Fix build, dependencies, NPE, and Ollama JSON compatibility
- Correct malformed pom.xml in eu.kalafatic.evolution.controller.
- Remove redundant version tag in eu.kalafatic.evolution.selfdev.genome/pom.xml.
- Update MANIFEST.MF to re-export eu.kalafatic.evolution.selfdev.genome.
- Add missing imports in EvolutionaryObservabilityManager.
- Implement robust null guard in MultiPageEditor.pageChange.
- Fix NoSuchMethodError in OllamaService by casting floats to doubles for JSONObject.put.
- Verified full project build with Maven.

+++ eu.kalafatic.evolution.controller/src/eu/kalafatic/evolution/controller/manager/OllamaService.java
@@ -179,7 +179,7 @@ public class OllamaService {
         jsonObject.put("prompt", prompt);
         jsonObject.put("stream", false);
         JSONObject options = new JSONObject();
-        options.put("temperature", this.temperature);
+        options.put("temperature", (double) this.temperature);
         options.put("num_predict", this.numPredict);
         jsonObject.put("options", options);
 
@@ -206,11 +206,11 @@ public class OllamaService {
         json.put("stream", stream);
 
         JSONObject options = new JSONObject();
-        options.put("temperature", this.temperature);
+        options.put("temperature", (double) this.temperature);
         options.put("num_predict", this.numPredict);
-        options.put("top_p", this.topP);
+        options.put("top_p", (double) this.topP);
         options.put("top_k", this.topK);
-        options.put("repeat_penalty", this.repeatPenalty);
+        options.put("repeat_penalty", (double) this.repeatPenalty);
         json.put("options", options);
 
         JSONArray msgs = new JSONArray();
